pipelines/video-dataset/requirements.txt
av>=10.0.0
numpy>=1.24.0
pillow>=10.0.0
librosa>=0.10.0
soundfile>=0.12.0
tqdm>=4.65.0
datasets>=2.14.0
pyyaml>=6.0
requests>=2.31.0
boto3>=1.28.0
pytest>=7.4.0
transformers>=4.35.0
openai-whisper>=20231117
accelerate>=0.24.0
encodec>=0.1.1
torch>=2.0.0
torchaudio>=2.0.0
openai>=1.0.0
